requirements.txt
Appium-Python-Client==3.1.1
atomicwrites==1.4.1
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest==6.2.5
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
